pyproject.toml
[tool.poetry]
name = "nhs-api-oauth2-client"
version = "0.1.0"
description = ""
authors = ["Thomas Judd-Cooper <thomas.judd-cooper1@nhs.net>"]
readme = "README.md"
packages = [{include = "nhs_api_oauth2_client"}]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.3.0"
requests = "^2.31.0"
pytest = "^7.4.0"
freezegun = "^1.2.2"
pyjwt = "^2.8.0"
cryptography = "^41.0.3"
requests-mock = "^1.11.0"
parameterized = "^0.9.0"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
isort = "^5.12.0"
pylint = "^2.17.5"
pytest-cov = "^4.1.0"

[tool.pytest.ini_options]
minversion = "7.4"
addopts = "-ra -q --cov=nhs_api_oauth2_client --cov-report term-missing"
testpaths = [
    "tests"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
